Folder/requirements.txt
pandas
seaborn
numpy
matplotlib
scikit-learn
tensorflow
opencv-python
dill
flask
pillow

-e .